+++ gala/potential/potential/__init__.py
@@ -4,3 +4,4 @@ from .ccompositepotential import *
 from .builtin import *
 from .io import *
 from .util import *
+from .interop import *